+++ TaskChecker/TaskChecker/GuiControl/ProcessListView/ProcessListView.Event.cs
@@ -13,22 +13,22 @@ namespace TaskChecker.GuiControl
         private void _addToolButton_Click( object pSender, EventArgs pEventArgs )
         {
             AddProcessItem(new TaskListItem.Entity {
-                id               = _children.Count,
+                id               = _processListLayoutPanel.childrenCount,
                 isEnableMemoArea = false,
                 isExpanded       = false,
                 processState     = TaskState.NOT_WORKING,
                 onClickSelected  = value => { SetChildSelected(value.id, true); },
                 onResizeRequest  = ( pItem, pSize ) =>
                 {
-                    if ( pItem.id >= _children.Count ) { pItem.Size = pSize; return; }
+                    if ( pItem.id >= _processListLayoutPanel.childrenCount ) { pItem.Size = pSize; return; }
                     
-                    Size fNewLayoutPanelSize = _listItemLayoutPanel.Size;//新たに設定するLayoutPanelのサイズ
+                    Size fNewLayoutPanelSize = _processListLayoutPanel.layoutPanelSize;//新たに設定するLayoutPanelのサイズ
 
                     fNewLayoutPanelSize.Height -= pItem.Size.Height;
                     fNewLayoutPanelSize.Height += pSize.Height;
                     
-                    _listItemLayoutPanel.RowStyles[pItem.id].Height = pSize.Height;
-                    _listItemLayoutPanel.Size = fNewLayoutPanelSize;
+                    _processListLayoutPanel.layoutPanelLineStyles[pItem.id].Height = pSize.Height;
+                    _processListLayoutPanel.layoutPanelSize = fNewLayoutPanelSize;
                     
                     pItem.Size = pSize;
                 },
@@ -40,7 +40,7 @@ namespace TaskChecker.GuiControl
         /// </summary>
         private void _removeToolButton_Click( object pSender, EventArgs pEventArgs )
         {
-            if ( _selections.Count <= 0 ) { return; }
+            if ( _processListLayoutPanel.selectionCount <= 0 ) { return; }
 			
             DialogResult fResult = MessageBox.Show("選択された作業工程を削除します。\nよろしいですか？", "削除確認", MessageBoxButtons.YesNo);
